Added support for Ubuntu Core image download during installation (for development purposes) To use it: - set USE_LOCAL_SOURCE to false in src/.config - comment IMAGE_URL in src/minimal_overlay/bundles/image_deploy/.config - replace ip and port of the web server and the file name of the uc image in src/minimal_overlay/bundles/image_deploy/99_deploy_uc20.sh (default is http://192.168.1.100:9000/pc.img.xz) - build minimal and flash on a usb stick as usual - start a web server on the directory containing your uc image (e.g. python3 -m http.server --bind 192.168.1.100 9000) - boot the device with minimal usb stick as usual

+++ src/minimal_overlay/bundles/image_deploy/bundle.sh
@@ -8,8 +8,9 @@ set -e
 mkdir -p $DEST_DIR/etc/autorun
 mkdir -p $DEST_DIR/root
 cp $SRC_DIR/99_deploy_uc20.sh $DEST_DIR/etc/autorun
-cp $OVERLAY_SOURCE_DIR/ubuntu-core-20-amd64.img.xz $DEST_DIR/root
-
+if [[ -e /root/ubuntu-core-20-amd64.img.xz ]]; then
+	cp $OVERLAY_SOURCE_DIR/ubuntu-core-20-amd64.img.xz $DEST_DIR/root
+fi
 
 chmod +x $DEST_DIR/etc/autorun/99_deploy_uc20.sh
